+++ src/Properties/FloatProperty.php
@@ -2,13 +2,13 @@
 
 namespace Elsevier\JSONSchemaPHPGenerator\Properties;
 
-class IntegerProperty extends ScalarProperty
+class FloatProperty extends ScalarProperty
 {
     /**
      * @param string $name
      */
     public function __construct($name)
     {
-        parent::__construct($name, 'integer');
+        parent::__construct($name, 'float');
     }
 }