+++ tests/CodeCreatorTest.php
@@ -6,7 +6,7 @@ use Elsevier\JSONSchemaPHPGenerator\CodeCreator;
 
 class CodeCreatorTest extends \PHPUnit\Framework\TestCase
 {
-    public function testCreatesClassWithIntegerProperty()
+    public function testCreatesClassWithNumberProperty()
     {
         $schema = json_decode('{
             "properties": {
@@ -16,12 +16,12 @@ class CodeCreatorTest extends \PHPUnit\Framework\TestCase
                 "foo"
             ]
         }');
-        $codeCreator = new CodeCreator('IntegerProperty', 'Elsevier\JSONSchemaPHPGenerator\Examples');
+        $codeCreator = new CodeCreator('FloatProperty', 'Elsevier\JSONSchemaPHPGenerator\Examples');
 
         $code = $codeCreator->create($schema);
 
         assertThat($code, arrayWithSize(1));
-        assertThat($code, hasClassThatMatchesTheExample('IntegerProperty'));
+        assertThat($code, hasClassThatMatchesTheExample('FloatProperty'));
     }
     
     public function testCreatesClassWithStringProperty()
@@ -290,5 +290,35 @@ class CodeCreatorTest extends \PHPUnit\Framework\TestCase
 
         assertThat($code, arrayWithSize(atLeast(1)));
         assertThat($code, hasClassThatMatchesTheExample('SubReferenceEnumFoo'));
+    }
+
+    public function testWithArrayOfReferencedObjects()
+    {
+        $schema = json_decode('{
+            "properties": {
+                "bar": {
+                    "items": {
+                        "$ref": "#/definitions/SubReference"
+                    },
+                    "type": "array"
+                },
+                "foo": {
+                    "items": {
+                        "$ref": "#/definitions/SubReference"
+                    },
+                    "type": "array"
+                }
+            },
+            "type": "object",
+            "required": [
+                "bar"
+            ]
+        }');
+        $codeCreator = new CodeCreator('ArrayOfObjects', 'Elsevier\JSONSchemaPHPGenerator\Examples');
+
+        $code = $codeCreator->create($schema);
+
+        assertThat($code, arrayWithSize(atLeast(1)));
+        assertThat($code, hasClassThatMatchesTheExample('ArrayOfObjects'));
     }
 }